+++ src/interfaces.ts
@@ -12,3 +12,14 @@ export interface Sprites {
 	content: string;
 	svgs: Array<string>;
 }
+
+export interface NormalModule {
+	userRequest: string;
+	originalSource: Function;
+}
+
+export interface Chunk {
+	buildInfo: Object;
+	userRequest: string;
+	originalSource: Function;
+}